Rebase against cluster-state-improvements
